refactor phase 3 - name changes, tidying up, restructuring
