+++ src/forms/DragHandler.js
@@ -1,0 +1,20 @@
+import { formDragEnded } from 'forms/FormActions';
+
+const DragHandler = {
+  beginDrag(props, monitor, component) {
+    return {
+      field: props.field,
+      id: props.id,
+      onList: props.onList,
+      position: props.position,
+      component: component
+    };
+  },
+  endDrag(props, monitor, component) {
+    // dispatchProps should be merged with props but looks like it's not inside this call
+    if (component && component.dispatchProps) component.dispatchProps.dispatch(formDragEnded());
+    if (props.formDiagram) props.formDiagram.setState({ isHovering: false });
+  }
+};
+
+export default DragHandler;
